Keep dota dataset devkit

+++ utils/nms/cpu_nms.pyx
@@ -154,7 +154,7 @@ def cpu_nms(
     cdef np.ndarray[np.intp_t, ndim=1] order = scores.argsort()[::-1]
 
     cdef int ndets = dets.shape[0]
-    cdef np.ndarray[np.int_t, ndim=1] suppressed = np.zeros((ndets), dtype=np.int)
+    cdef np.ndarray[np.int_t, ndim=1] suppressed = np.zeros((ndets), dtype=np.int32)
 
     cdef int _i, _j, i, j, rtn
     cdef np.float32_t inter, ovr
